+++ packages/seedless-onboarding-controller/src/utils.ts
@@ -15,6 +15,7 @@ import type {
   DecodedNodeAuthToken,
   DeserializedVaultData,
   InvalidPrimarySecretDataTypeErrorData,
+  SentryError,
   VaultData,
 } from './types';
 
@@ -199,3 +200,29 @@ export function summarizeSecretMetadataCounts(
 
   return { secretTypeCounts, dataTypeCounts };
 }
+
+/**
+ * Creates a Sentry error from an error message, an inner error and a context.
+ *
+ * NOTE: Sentry defaults to a depth of 3 when extracting non-native attributes.
+ * As such, the context depth shouldn't be too deep.
+ *
+ * @param message - The error message to create a Sentry error from.
+ * @param innerError - The inner error to create a Sentry error from.
+ * @param context - The context to add to the Sentry error.
+ * @returns A Sentry error.
+ */
+export const createSentryError = <
+  TContext extends Record<string, unknown> = Record<string, unknown>,
+>(
+  message: string,
+  innerError: Error,
+  context?: TContext,
+): SentryError<TContext> => {
+  const error = new Error(message) as SentryError<TContext>;
+  error.cause = innerError;
+  if (context) {
+    error.context = context;
+  }
+  return error;
+};